Refactor models to remove unused HasFactory trait and update relationships. Update migration files to change cost fields to bigInteger. Remove factory files for Customer, Driver, Fleet, FleetType, IndustrySector, User, and Vendor. Add detailed customer seeder data. Update tests to create users and handle authentication directly.

+++ database/migrations/2025_08_12_080608_create_expedition_costs_table.php
@@ -14,31 +14,31 @@ return new class extends Migration
         Schema::create('expedition_cost_fleets', function (Blueprint $table) {
             $table->id();
             $table->foreignId('expedition_id')->constrained('expeditions');
-            $table->integer('sales_amount')->nullable();
-            $table->integer('deposit_cost')->nullable();
+            $table->bigInteger('sales_amount')->nullable();
+            $table->bigInteger('deposit_cost')->nullable();
             $table->date('deposit_date')->nullable();
-            $table->integer('transportation_cost')->nullable();
-            $table->integer('fuel_cost')->nullable();
-            $table->integer('tolling_cost')->nullable();
-            $table->integer('port_cost')->nullable();
-            $table->integer('insurance_cost')->nullable();
-            $table->integer('driver_cost')->nullable();
-            $table->integer('other_cost')->nullable();
+            $table->bigInteger('transportation_cost')->nullable();
+            $table->bigInteger('fuel_cost')->nullable();
+            $table->bigInteger('tolling_cost')->nullable();
+            $table->bigInteger('port_cost')->nullable();
+            $table->bigInteger('insurance_cost')->nullable();
+            $table->bigInteger('driver_cost')->nullable();
+            $table->bigInteger('other_cost')->nullable();
             $table->string('description')->nullable();
-            $table->integer('total_cost')->nullable();
+            $table->bigInteger('total_cost')->nullable();
             $table->timestamps();
         });
 
         Schema::create('expedition_cost_vendors', function (Blueprint $table) {
             $table->id();
             $table->foreignId('expedition_id')->constrained('expeditions');
-            $table->integer('sales_amount')->nullable();
-            $table->integer('vendor_cost')->nullable();
-            $table->integer('deposit_cost')->nullable();
+            $table->bigInteger('sales_amount')->nullable();
+            $table->bigInteger('vendor_cost')->nullable();
+            $table->bigInteger('deposit_cost')->nullable();
             $table->date('deposit_date')->nullable();
-            $table->integer('other_cost')->nullable();
+            $table->bigInteger('other_cost')->nullable();
             $table->string('description')->nullable();
-            $table->integer('total_cost')->nullable();
+            $table->bigInteger('total_cost')->nullable();
             $table->timestamps();
         });
     }
